Removed useless package
