+++ packages/boot-contract-derive/Cargo.toml
@@ -10,6 +10,10 @@ description = "Attribute macro for creating a contract interface."
 [lib]
 proc-macro = true
 
+[features]
+default=["outside_contract"]
+outside_contract=[]
+
 [dependencies]
 convert_case = "0.6.0"
 cosmwasm-std.workspace = true
