--- packages/boot-macros/Cargo.toml
@@ -1,19 +1,0 @@
-[package]
-name = "boot-macros"
-version = "0.10.0"
-authors = { workspace = true}
-edition = { workspace = true}
-license = { workspace = true}
-repository = { workspace = true}
-description = "Attribute macro behind a boot flag for easier importing"
-
-[lib]
-proc-macro = true
-
-[features]
-outside_contract=["dep:boot-contract-derive"]
-
-[dependencies]
-boot-contract-derive = { git = "https://github.com/Kayanski/BOOT", branch = "feature/new-contract-interface", optional=true}
-quote = "1.0.26"
-syn = { version = "1", features = ["full", "extra-traits"] }
